Refactor CGroup
